requirements.txt
paho-mqtt==1.6.1
fpdf2==2.7.7
pyshark==0.6
